Close the city view when city overview panel being closed due to selection change (#12)
When the selection changes (i.e. buying building with gold or starting a
new building in a district), the city overview panel closes in order to
show some on-map animation. However, the production panel remains open
and the view remains zoomed in. This commit fixes the issue (tracked by
bug #10).