Reimplemented the cityview!
I essentially gutted the last implementation and rewrote it from the
ground up. This implementation is more stateful and can tolerate much more
abuse. This implementation also resolves issues #10 and #19, along with
providing a fix for the Civ VI bug requiring the city ranged strike
button to be pressed twice. Yay for bugfixes!