Adjusted citizen management icon transparency
The citizen management icons are now semi-transparent. This is a change
designed to make tile yields more readable within the cityview. In order
to preserve the usability of the citizen icons, the button size now also changes
in addition to the enable/disable state of the texture. This should make
it easier to tell if a tile is being worked or not, despite the change to
button transparency