Adds Hover information when mousing over City Banner Resolves #136
When mousing over a city, the yields for that city are displayed. Tiles
being worked are represented using the same graphics as when one is
managing citizens. The next culture plot is displayed using the CQUI
next plot button. Also in this commit is making the "y" hotkey do the
same action as pressing on the "Show Yields" checkbox in the minimap map
options.
TODO: Display colored culture hex
TODO: More efficient way to only reveal yields for a specific city