Updated contributors
It's been way too long since I've done this! Huge thanks to @jacks0nX, @zeyangl, @RatchetJ, @velit, @astog, @the-m4a,
and @ZhouYzzz for their excellent work on CQUI! I cannot aqequately
express my appreciation for all of the consistently high quality work the
CQUI contributors bring to the table, you folks are the lifeblood and
magic of the open source community and the reason CQUI is as good as it
is. :heart: