Refactored CityOverviewPanel bubbles
Lots of background changes to the citypaneloverview bubbles. They look
worse now, but they're no longer incredibly easy to break nor impossible
to localize! This was the last place custom LOCs could not be inserted
and now that's resolved once and for all.

This should also pave the way for making these icons more customizable.
For now, the only new functionality is that the icons are properly
color-coded. In the future, expect to also have the option to only show
bubbles with an active modifier