Added ToString support to data_converter handling
This commit adds the option to include a ToString function in the
data_converter that gets passed to slider controls. This function
provides instructions for formatting the output of ToValue for
displaying in the label. When none is provided, the output of ToValue is
used for the label, just like before