Fixes #301: Linux compatibility
This commit changes every single file to lowercase. Why? Because Linux
is case sensitive and all of the files in the linux version of the game
are lowercase for some reason. Not sure why the file names seem to be
deliberately different case *just* for linux, but since it's the only
platform this actually matters on, the new convention is that all mod
files are lowercase.

Anyways, all of my linux machines are headless, so I can't test to see
if this actually fixes the Linux issue, I'm going to need someone to go
out and test these changes to make sure!