Fixes promotion symbol
Finally fixes the promotion symbol so that it is always displayed for local player units when there is a promotion action available. Previously, due to faulty Firaxis functions regarding XPs (not updating between promotion actions), the symbol would stay put even if the promotion action was taken. Also, previous code always overwrote the promotion symbol when there were already promotions for the unit. Now, promotion symbol is displayed even if there are already other promotions for the unit, and changes to the new number of promotions immediately after the promotion action.