Fixed a possible crash post-BulkHide
This is a theoretical fix for an issue which I've been unable to
reproduce. For some reason, TeamRibbon does not seem to load up after a
BulkHide for some users, this may be by design or a bug in the core
game, either way, now our code checks for this issue and sidesteps it
now.