Real Housing - recalculate values only when they changed
Previously real housing from improvements function was updating data between turns or when changing city focus or changing government and so on. And as soon as housing from improvements never changes on these actions now we don't recalculate the same values and take them from a table. We recalculate these values when tiles improved/pillaged, when swap tiles and when tiles lost by culture bomb only. This should also improve performance a bit.