Real Housing Fall 2017 update and improve
Added Indonesian Kampungs to real housing from improvements code. Also improved data update when a city loses a tile to a Culture Bomb or when research Sanitation when play for India. Also added a function to update data for a city that loses a 3rd radius tile to a city that is founded within 4 tiles.