Revert unnecessary update citizen and citybanner rightclick
'CQUI_UpdateSelectedCityCitizens' function was added to update cities data when swap or improve tiles. Now we update it without this function so we don't need it. It also freezes the camera every time when choose another city from a cityview. So here we remove it.
'OnCityBannerRightClick' function was added to recenter camera on city. It makes harder to move units to plots that are adjacent to city centers sometimes. We have a Middle Click for recentering camera anyway so here we remove this function also.