Adds gold balance and yield to leader icon tooltip
This patch add the gold balance and yield to each leader icon tooltip in the diplo ribbon. Vanilla civ 6 does not display information when the yield of a given leader is negative (inside the deal screen), so we follow the same pattern to preserve the non-gameplay-altering nature of CQUI. When yield is negative, tooltip displays a ? symbol.