reworked code
Reworked main function to calculate real housing a bit faster. Also added some conditions to 'citypanel' and 'OnTileImproved' function to recalculate housing only for improvements or when they are removed and to don't recalculate it when remove features or resources for examle. Reworked swap tiles data and housing update code to work on 'CityTileOwnershipChanged' event. Previously it could try to recalculate data before this event.